Refactor, add other reports
