+++ app/services/import_notifier.rb
@@ -10,31 +10,33 @@ class ImportNotifier
   # data feed readings, creating in last 24 hours, where the readings are ALL blank
   # nil, numeric, string values are not blank, so this equates to an array of ['']
   def meters_with_blank_data(from: 24.hours.ago, to: Time.zone.now)
-    meters = Meter.active
+    Meter.active
     .joins(:school)
     .joins(:amr_data_feed_readings)
-    .where(schools: { active: true })
-    .where('amr_data_feed_readings.readings = ARRAY[?]', Array.new(48, '')) # where readings is empty string
     .joins('INNER JOIN amr_data_feed_import_logs on amr_data_feed_readings.amr_data_feed_import_log_id = amr_data_feed_import_logs.id') # manually join to import logs
+    .where(schools: { active: true })
     .where('import_time BETWEEN :from AND :to', from: from, to: to) # limit to period
+    .where('amr_data_feed_readings.readings = ARRAY[?]', Array.new(48, '')) # where readings is empty string
+    .includes(:school, { school: :school_group }, :procurement_route, :data_source, :admin_meter_status, :issues)
     .distinct # distinct meters
-    meters.sort_by {|m| [m.school.area_name, m.meter_type, m.school_name, m.mpan_mprn]}
+    .order({ school_groups: { name: :asc } }, :meter_type, { schools: { name: :asc } }, :mpan_mprn)
   end
 
   # data feed readings, creating in last 24 hours, where the readings are ALL 0 or 0.0
   # this version is slightly different to original as that used ruby to cast values to a float
   # this meant any dodgy chars, e.g. '-', where treated as 0.0
   def meters_with_zero_data(from: 24.hours.ago, to: Time.zone.now)
-    meters = Meter.active
+    Meter.active
     .joins(:school)
+    .joins(:amr_data_feed_readings)
+    .joins('INNER JOIN amr_data_feed_import_logs on amr_data_feed_readings.amr_data_feed_import_log_id = amr_data_feed_import_logs.id') # manually join to import logs
     .where.not(meter_type: :exported_solar_pv) # exported solar PV is legitimately zero on some days
     .where(schools: { active: true })
-    .joins(:amr_data_feed_readings)
-    .where('amr_data_feed_readings.readings = ARRAY[?] OR amr_data_feed_readings.readings = ARRAY[?]', Array.new(48, '0'), Array.new(48, '0.0')) # where readings are 0, or 0.0
-    .joins('INNER JOIN amr_data_feed_import_logs on amr_data_feed_readings.amr_data_feed_import_log_id = amr_data_feed_import_logs.id') # manually join to import logs
     .where('import_time BETWEEN :from AND :to', from: from, to: to) # limit to period
+    .where('amr_data_feed_readings.readings = ARRAY[?] OR amr_data_feed_readings.readings = ARRAY[?]', Array.new(48, '0'), Array.new(48, '0.0')) # where readings are 0, or 0.0
+    .includes(:school, { school: :school_group }, :procurement_route, :data_source, :admin_meter_status, :issues)
     .distinct # distinct meters
-    meters.sort_by {|m| [m.school.area_name, m.meter_type, m.school_name, m.mpan_mprn]}
+    .order({ school_groups: { name: :asc } }, :meter_type, { schools: { name: :asc } }, :mpan_mprn)
   end
 
   def notify(from:, to:)
