+++ app/views/admin/reports/base_meter_reports/_report_filters.html.erb
