+++ app/views/admin/reports/base_meter_reports/index.html.erb
