+++ app/views/admin/reports/index.html.erb
@@ -63,6 +63,14 @@
     <h3>Data Imports</h3>
     <ul>
       <li><%= link_to 'Data feed import logs', admin_reports_amr_data_feed_import_logs_path %></li>
+      <li>
+        Meters with:
+         <ul>
+           <li><%= link_to 'Blank readings', admin_reports_blank_readings_path %></li>
+           <li><%= link_to 'Stale data', admin_reports_lagging_meters_path %></li>
+           <li><%= link_to 'Zero readings', admin_reports_zero_readings_path %></li>
+         </ul>
+      </li>
       <li><%= link_to 'Meter loading report', admin_reports_meter_loading_reports_path %></li>
       <li><%= link_to 'Recent manual imports', admin_reports_data_loads_path %></li>
       <li><%= link_to 'School group meter reports', admin_reports_meter_reports_path %></li>
